src/main/java/io/github/korzepadawid/springtaskplanning/security/CustomUserDetailsService.java
package io.github.korzepadawid.springtaskplanning.security;

import io.github.korzepadawid.springtaskplanning.repository.UserRepository;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class CustomUserDetailsService implements UserDetailsService {

  private final UserRepository userRepository;

  public CustomUserDetailsService(UserRepository userRepository) {
    this.userRepository = userRepository;
  }

  @Override
  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
    return userRepository
        .findByEmail(username)
        .map(UserPrincipal::create)
        .orElseThrow(
            () -> new UsernameNotFoundException(String.format("User %s doesn't exist", username)));
  }
}
